Update the configuration bits
FCMEM was presumably a typo of FCMEN.

A brown-out reset voltage (BORV) of 21 is not valid
in mplabc18-v3.40. It has been replaced with a valid value
of 3. We don't care about this value and could probably remove
it because brown-out reset (BOR) is set to "OFF".

The microcontroller this project targets does not support
ICPRT. This commit removes the config option disabling it.